Movie Store
Assignment 4 AKA Movie Store was an assignment that I worked on with my partner Jeff Taylor. The Movie Store assignment was a demonstration of my cumulative knowledge obtained through the first two data structures classes at the University of Washington Bothell. This program showed our knowledge of inheritance, hash tables, and trees. A large assignment, the Movie Store program is a backend inventory management system that would be easy to update and scale with a growing number of products to be sold or rented.