Soccer field line detection
This code is to be used on the raspberry pi for the elcano autonomous trike. The code uses the Raspi camera and isolates for yellow lines, getting the distance from the bottom of the detected line and the angle of the line to communicate to to the high level board which sends instructions to the motor of the trike to turn.